requirements.txt
numpy
mediapipe
opencv-python
python-dotenv
pymongo
rich
asyncio
websockets